requirements.txt
# praw==7.8.1
# python-dotenv==1.0.0
# beautifulsoup4==4.12.2
# requests==2.31.0
# google-generativeai==0.3.2

# praw==7.8.1
# python-dotenv==1.0.0
# beautifulsoup4==4.12.2
# requests==2.31.0
# deepseek

# huggingface_hub>=0.20.0
# requests>=2.31.0
# python-dotenv>=1.0.0


Flask
praw
requests
beautifulsoup4
python-dotenv
fpdf2
Pillow
together
huggingface_hub
imgkit
pdfkit
pyppeteer